Expand debug.sh so I can avoid launching OpenOCD everytime

+++ debug.sh
@@ -3,8 +3,15 @@ set -e
 cmake --preset Debug
 cd build/Debug
 ninja
-openocd -f interface/stlink.cfg -f target/stm32g4x.cfg -c "program ECU-25.elf" &
-OCDPID=$!
-unset -e
-gdb --symbols=ECU-25.elf --eval-command="target extended-remote :3333"
-kill $OCDPID
+exitcode="${?}"
+
+if [ "$1" != "build" ] && [ "$1" != "b" ]
+then
+	openocd -f interface/stlink.cfg -f target/stm32g4x.cfg -c "program ECU-25.elf" &
+	OCDPID=$!
+	unset -e
+	gdb --symbols=ECU-25.elf --eval-command="target extended-remote :3333"
+	kill $OCDPID
+else
+	echo "Ninja exited with status ${exitcode}"
+fi
